REF: move SelectN from core.algorithms
